% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_excerpts.R
\name{view_excerpts_table}
\alias{view_excerpts_table}
\title{Interactive excerpts table}
\usage{
view_excerpts_table(data)
}
\arguments{
\item{data}{A data.frame or tibble with one column called \code{excerpt}
and multiple code columns starting with \code{"c_"}.}
}
\value{
A DT datatable widget.
}
\description{
Creates an interactive datatable where users can filter excerpts by code.
Automatically detects all code columns starting with "c_".
}
\examples{
library(dplyr)
library(tidyr)
library(DT)

df <- tibble::tibble(
  excerpt = c(
    "I felt connected to peers.",
    "We should normalize conversations about MH.",
    "My teachers helped me belong.",
    "I am comfortable talking about suicide."
  ),
  c_belonging = c(TRUE, FALSE, TRUE, FALSE),
  c_destigmatization = c(FALSE, TRUE, FALSE, FALSE),
  c_suicide_comfort = c(FALSE, FALSE, FALSE, TRUE)
)

view_excerpts_table(df)

}
