% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_saturation.R
\name{plot_saturation}
\alias{plot_saturation}
\title{Plot Saturation Tracking Bar Chart}
\usage{
plot_saturation(long_codes)
}
\arguments{
\item{long_codes}{A data frame output from create_saturation_tracking()}
}
\value{
A ggplot object showing stacked bar chart of Priority and
Heterogeneity code counts
}
\description{
Plot Saturation Tracking Bar Chart
}
